+++ javascript/src/types/index.ts
@@ -2,7 +2,6 @@ export * from "./common.js";
 export * from "./input.js";
 export * from "./tools.js";
 export * from "./responses.js";
-export * from "./safety-identifiers.js";
 export * from "./streaming.js";
 export * from "./catalog.js";
 export * from "./volumes.js";